fix(swap-cryptos): removed classes file, all moved to swap-crypto-utils
